vsp2-renderer: use model name to determine the underlying device.
We used to refer bus_info to determine the underlying VSP device.
Instead, the model name is the information we should refer to determine
the devie.

However, as of Linux Kernel 4.6.0, the model name doesn't allow us to
distinguish between VSPI and VSPB. This needs to be fixed in the future
in the kernel driver.

Reported-by: Tomohito Esaki <etom@igel.co.jp>