vsp2-renderer: change behavior of scaling and surface rotation.
When gl-fallback is enabled, surface rotation is supported with GPU.

Scaling is also either by VSP, or by GPU if gl-fallback is enabled.
If scaling with VSP and gl-fallback are enabled, VSP does scaling
in the case without rotation, and GPU does scaling + rotation in the
case with both rotation and scaling.

The name of a flag to enable VSP scaling has also been changed as
follows:

[vsp-renderer]
vsp-scaler=true