v4l2-renderer: use stride of shm buffer, not of KMS BO
KMS BO allocates buffer that meets the constraint of underlying
hardware. However V4L2 may not have the same constraint, and often
doesn't.

For instance, the KMS BO may allocates the buffer with the stride in
unit of 64 bytes. We should ignore the stride that KMS BO came up with,
and use the number we got from shm buffer to meet the requirement in
V4L2.